new module for logging HTTP connections

+++ handlers/signal_handler.py
@@ -5,7 +5,11 @@ import sys
 
 def signal_handler(signal,frame):
     print "Request exit"
-    shutil.rmtree(os.getcwd()+"/certs")
+    if(os.path.exists(os.getcwd()+"/certs")):
+        # If we did not see any TLS/SSL connections, then this folder will not exist
+        shutil.rmtree(os.getcwd()+"/certs")
+    
+    shutil.rmtree(os.getcwd()+"/http_logging")
     sys.exit(0)
